refactor: split config tests into src/config/tests.rs
config.rs was 825 lines, ~47% an inline #[cfg(test)] mod tests. Move the tests
verbatim to a child submodule at src/config/tests.rs (use super::*, full
crate-internal access retained since tests stays a child of the config module).

Modern no-mod.rs layout: config.rs stays the module, config/ holds its test
submodule. Public API unchanged. Pure extraction, no logic change; cargo test
runs the identical 25-test set.
